The MarkdownTransformer will do two things: parse markdown and eventually cache it. Let's start with the first.
Open up GenusController and copy the code that originally parsed the text through markdown. Now... paste this into the parse function, return that and update the variable to {dollar_sign}str. Wow, that was easy.

Go back and refresh. It explodes!

Attempted to call an undefined method named 'get' on MarkdownTransformer
Forget about Symfony: this makes sense. The class we just created does not have a get() function and it doesn't extend anything that would give this to us.

In a controller, we do have this function. But more importantly, we have access to the container, either via that shortcut method or by saying {dollar_sign}this->container. From there we can fetch any service by calling ->get() on it.

But that's special to the controller: as soon as you're not in a controller - like MarkdownTransformer - you don't have access to the container, its services... or anything.