What type of object is this  argument? Oh, you can't tell? Well, neither can I. With no type-hint, this could be anything! A MarkdownParser object, a string, an octopus!
We need to add a typehint to make our code clearer... and avoid weird errors in case we accidentally pass in something else... like an octopus.

Run:

./bin/console debug:container markdown
And select markdown.parser - that's the service we're passing into MarkdownTransformer. Ok, it's an instance of Knp\Bundle\MarkdownBundle\Parser\Preset\Max. We can use that as the type-hint.

But hold on - I'm going to complicate things... but then we'll all learn something cool and celebrate. Press shift+shift, type 'max' and open that class:
14 lines  vendor/knplabs/knp-markdown-bundle/Parser/Preset/Max.php
... lines 1 - 6
/**
 * Full featured Markdown Parser
 */
class Max extends MarkdownParser {
}

Ah, this extends MarkdownParser and that does all the work:
245 lines  vendor/knplabs/knp-markdown-bundle/Parser/MarkdownParser.php
... lines 1 - 8
/**
 * MarkdownParser
 *
 * This class extends the original Markdown parser.
 * It allows to disable unwanted features to increase performances.
 */
class MarkdownParser extends MarkdownExtra implements MarkdownParserInterface {
... lines 17 - 243
}

And this implements a MarkdownParserInterface. We could type-hint with Max, MarkdownParser or MarkdownParserInterface: they will all work. BUT, when possible, it's best to find a base class - or better - and interface that has the methods on it you need, and use that.

Type-hint the argument with MarkdownParserInterface:
22 lines  src/AppBundle/Service/MarkdownTransformer.php
... lines 1 - 4
use Knp\Bundle\MarkdownBundle\MarkdownParserInterface;
class MarkdownTransformer {
... lines 9 - 10
  public function __construct(MarkdownParserInterface {dollar_sign}markdownParser)
... lines 12 - 20
}

Why is this the best option? Two small reasons. First, in theory, we could swap out the {dollar_sign}markdownParser for a different object, as long as it implemented this interface. Second, it's really clear what methods we can call on the {dollar_sign}markdownParser property: only those on that interface.

But hold on a second, PhpStorm is angry about calling transform() on {dollar_sign}this->markdownParser:

Method 'transform' not found in class MarkdownParserInterface
Weird! Open that interface. Oh, it has only one method: transformMarkdown():
16 lines  vendor/knplabs/knp-markdown-bundle/MarkdownParserInterface.php
... lines 1 - 4
interface MarkdownParserInterface {
  /**
   * Converts text to html using markdown rules
   *
   * @param string  plain text
   *
   * @return string rendered html
   */
  function transformMarkdown({dollar_sign}text);
}
Hold on: to be clear: everything will work right now. Refresh to prove it.

The weirdness is just that we are forcing an object that implements MarkdownParserInterface to be passed in... but then we're calling a method that's not on that interface.