The MarkdownTransformer class is not an Autobot, nor a Decepticon, but it is a service because it does work for us. It's no different from any other service, like the markdown.parser, logger or anything else we see in debug:container... except for one thing: it sees dead people. I mean, it does not live in the container.
Nope, we need to instantiate it manually: we can't just say something like {dollar_sign}this->get('app.markdown_transformer') and expect the container to create it for us.

Time to change that... and I'll tell you why it's awesome once we're done.

Open up app/config/services.yml. To add a new service to the container, you basically need to teach the container how to instantiate your object.

This file already has some example code to do this... kill it! Under the services key, give your new service a nickname: how about app.markdown_transformer. This can be anything: we'll use it later to fetch the service. Next, in order for the container to be able to instantiate this, it needs to know two things: the class name and what arguments to pass to the constructor. For the first, add class: then the full AppBundle\Service\MarkdownTransformer.
For the second, add arguments: then make a YAML array: []. These are the constructor arguments, and it's pretty simple. If we used [sunshine, rainbows], it would pass the string sunshine as the first argument to MarkdownTransformer and rainbow as the second. And that would be a very pleasant service.

In reality, MarkdowTransformer requires one argument: the markdown.parser service. To tell the container to pass that, add @markdown.parser.

That's it. The @ is special: it says:

Woh woh woh, don't pass the string markdown.parser, pass the service markdown.parser.
And with 4 lines of code, a new service has been born in the container. I'm such a proud parent.

Go look for it:

./bin/console debug:container markdown
There is its!