Instead of new MarkdownTransformer(), be lazier: {dollar_sign}transformer = {dollar_sign}this->get('app.markdown_transformer).
When this line runs, the container will create that object for us behind the scenes.

Why add a Service to the Container

Believe it or not, this was a huge step. When you add your service to the container, you get two great thing. First, using the service is so much easier: {dollar_sign}this->get('app.markdown_transformer). We don't need to worry about passing constructor arguments: heck it could have ten constructor arguments and this simple line would stay the same.

Second: if we ask for the app.markdown_transformer service more than once during a request, the container only creates one of them: it returns that same one object each time. That's nice for performance.

Oh, and by the way: the container doesn't create the MarkdownTransformer object until and unless somebody asks for it. That means that adding more services to your container does not slow things down.

The Dumped Container

Ok, I have to show you something cool. Open up the var/cache directory. If you don't see it - you may have it excluded: switch to the 'Project' mode in PhpStorm.

Open var/cache/dev/appDevDebugProjectContainer.php:

 3709 lines  var/cache/dev/appDevDebugProjectContainer.php
... lines 1 - 9
/**
 * appDevDebugProjectContainer.
 *
 * This class has been auto-generated
 * by the Symfony Dependency Injection Component.
 */
class appDevDebugProjectContainer extends Container {
... lines 18 - 3707
}
This is the container: it's a class that's dynamically built from our configuration. Search for 'MarkdownTransformer' and find the getApp_MarkdownTransformerService() method:

 3709 lines  var/cache/dev/appDevDebugProjectContainer.php
... lines 1 - 15
class appDevDebugProjectContainer extends Container {
... lines 18 - 23
  public function __construct(){
... lines 26 - 32
    {dollar_sign}this->methodMap = array(
... line 34
      'app.markdown_transformer' => 'getApp_MarkdownTransformerService',
... lines 36 - 238
    );
... lines 240 - 249
  }
... lines 251 - 272
  /**
   * Gets the 'app.markdown_transformer' service.
   *
   * This service is shared.
   * This method always returns the same instance of the service.
   *
   * @return \AppBundle\Service\MarkdownTransformer A AppBundle\Service\MarkdownTransformer instance.
   */
  protected function getApp_MarkdownTransformerService() {
    return {dollar_sign}this->services['app.markdown_transformer'] = new \AppBundle\Service\MarkdownTransformer({dollar_sign}this->get('markdown.parser'));
... lines 285 - 3707
}
Ultimately, when we ask for the app.markdown_transformer service, this method is called. And look! It runs the same PHP code that we had before in our controller: new MarkdownTransformer() and then {dollar_sign}this->get('markdown.parser') - since {dollar_sign}this is the container.

You don't need to understand how this works - but it's important to see this. The configuration we wrote in services.yml may feel like magic, but it's not: it causes Symfony to write plain PHP code that creates our service objects. There's no magic: we describe how to instantiate the object, and Symfony writes the PHP code to do that. This makes the container blazingly fast.