Phew! Dependency injection, check! Registering new services, check! Delicious snack, check! Well, I hope you just had a delicious snack.
This tutorial is the start to our victory lap. We need to add caching to MarkdownTransformer: it should be pretty easy. Copy part of the old caching code and paste that into the parse() function. Remove the else part of the if and just return {dollar_sign}cache->fetch().

Below, assign the method call to the  variable and go copy the old {dollar_sign}cache->save() line. Return {dollar_sign}str and re-add the sleep() call so that things are really slow - that keeps it interesting.

On top, change the  variables to . Perfect!

We know this won't work: there is no get() function in this class. And more importantly, we don't have access to the doctrine_cache.provider.my_markdown_cache service. How can we get access? Dependency injection.