Dependency Inject!
This time, add a second argument to the constructor called . And hmm, we should give this a type-hint. Copy the service name and run:

./bin/console debug:container doctrine_cache.providers.my_markdown_cache
This service is an instance of ArrayCache. But wait! Do not type-hint that. In our earlier course on environments, we setup a cool system that uses ArrayCache in the dev environment and FilesystemCache in prod:

 72 lines  app/config/config.yml
... lines 1 - 7
parameters:
  locale: en
  cache_type: file_system
... lines 11 - 65
doctrine_cache:
  providers:
    my_markdown_cache:
      type: %cache_type%
      file_system:
        directory: %kernel.cache_dir%/markdown_cache
If we type-hint with ArrayCache, this will explode in prod because this service will be a different class.

Let's do some digging: open up ArrayCache:

 95 lines  vendor/doctrine/cache/lib/Doctrine/Common/Cache/ArrayCache.php
... lines 1 - 32
class ArrayCache extends CacheProvider {
... lines 35 - 93
}
This extends CacheProvider:

 278 lines  vendor/doctrine/cache/lib/Doctrine/Common/Cache/CacheProvider.php
... lines 1 - 31
abstract class CacheProvider implements Cache, FlushableCache, ClearableCache, MultiGetCache {
... lines 34 - 276
}
That might work. But it implements several interface - one of them is just called Cache. Let's try that. If this isn't the right interface - meaning it doesn't contain the methods we're using - PhpStorm will keep highlighting those after we add the type-hint

I'll use a keyboard shortcut - option+enter on a Mac - and select initialize fields

All this did was add the private  property and set it in __construct(). You can also do that by hand.

Cool! Update parse() with {dollar_sign}cache = {dollar_sign}this->cache

And look! All of the warnings went away. That was the right interface to use. Yay!